Final Edit Cleaned up Code
The first thing is first you must drop the excel files into the python main folder. This program is designed for cleaning up both the master file and the excel export file. We use pandas for cleaning up, openpyxl to open the files, and DateTime to assign a date to the final file names. With this code, we are able to use the input to take in the user's input and put the file's name in which is saved as a variable.  That variable is then used in conjunction with the excel open method to assign the variable as the name of the file we are opening. It is important that the file you input is the exact same name of the file you dropped in the python main file.